Replace all use of gtest death testing with cmake death testing
GTest death testing is fundamentally broken. Even when it works,
it is very slow. This commit replaces GTest death testing with
cmake death testing, which is far faster and doesn't have any
of the quirks nor bugs. A new `monad_add_test_death` cmake
function makes cmake death testing available project wide.

This commit also removes `monad_target_gdb_pretty_printer`
which is no longer necessary now we have an IPC bridge to our
GDB fork.
