+++ libs/async/src/monad/async/test/CMakeLists.txt
@@ -12,7 +12,14 @@ add_async_test(TARGET cpp_coroutine_wrappers_test SOURCES
 
 add_async_test(TARGET io_test SOURCES "io.cpp")
 
-add_async_test(TARGET io_death_test SOURCES "io_death.cpp")
+monad_add_test_death(io_death_write_buffer_exhaustion_causes_death_test
+                     SOURCES "io_death_write_buffer_exhaustion_causes_death.cpp"
+                     LINK_LIBRARIES monad_async
+                     FAIL_REGEX ".*Must fail after this:\nFATAL: no i/o buffers remaining.*")
+monad_add_test_death(io_death_read_buffer_exhaustion_causes_death_test
+                     SOURCES "io_death_read_buffer_exhaustion_causes_death.cpp"
+                     LINK_LIBRARIES monad_async
+                     FAIL_REGEX ".*Must fail after this:\nFATAL: no i/o buffers remaining.*")
 
 add_async_test(TARGET io_worker_pool_test SOURCES "io_worker_pool.cpp")
 target_link_options(io_worker_pool_test PUBLIC "-rdynamic")
